+++ Enemy.java
@@ -6,12 +6,12 @@ public class Enemy extends Character{
 	public static HashMap<String, Enemy> dictionary = new HashMap<String, Enemy>();
 	
 	//Constructors
-	public Enemy(String name, double x, double y, int hp, int dmg, double width, double height) {
+	public Enemy(String name, int x, int y, int hp, int dmg, int width, int height) {
 		super(name, x, y, width, height, hp, dmg);
 		if(!dictionary.containsKey(name))
-			dictionary.put(name, new Enemy(name, 0.0, 0.0, hp, dmg, width, height));
+			dictionary.put(name, new Enemy(name, 0, 0, hp, dmg, width, height));
 	}
-	public Enemy(String name, double x, double y) {
+	public Enemy(String name, int x, int y) {
 		super(name,x,y,dictionary.get(name).width,dictionary.get(name).height,dictionary.get(name).hp,dictionary.get(name).dmg);
 	}
